+++ packages/server/drizzle/0017_dazzling_malice.sql
@@ -1,0 +1,7 @@
+CREATE TABLE `stripe_events` (
+	`event_id` text PRIMARY KEY NOT NULL,
+	`type` text NOT NULL,
+	`created_at` integer NOT NULL
+);
+--> statement-breakpoint
+ALTER TABLE `families` ADD `stripe_customer_id` text;